@@ -28,24 +28,21 @@
   },
   "dependencies": {
     "@date-io/moment": "^1.3.5",
-    "@material-ui/core": "^3.9.2",
-    "@material-ui/icons": "^3.0.2",
-    "bootstrap-4-grid": "^3.0.0",
+    "@fortawesome/fontawesome-svg-core": "^1.2.28",
+    "@fortawesome/free-solid-svg-icons": "^5.13.0",
+    "@fortawesome/react-fontawesome": "^0.1.9",
+    "bootstrap": "^4.4.1",
+    "chart.js": "^2.9.3",
     "classnames": "^2.2.6",
     "flag-icon-css": "^3.3.0",
+    "font-awesome": "^4.7.0",
     "leaflet": "^1.4.0",
-    "material-ui-pickers": "^2.2.4",
-    "material-ui-rating": "^3.0.1",
-    "material-ui-slider": "^3.0.8",
     "moment": "^2.24.0",
-    "normalize.css": "^8.0.1",
-    "pdfmake": "^0.1.56",
-    "react": "^16.5.0",
-    "react-contenteditable": "^3.3.1",
+    "react": "^16.13.1",
+    "react-bootstrap": "^1.0.0",
+    "react-chartjs-2": "^2.9.0",
     "react-datepicker": "^2.5.0",
     "react-dom": "^16.5.0",
-    "react-input-mask": "^2.0.4",
-    "react-leaflet": "^2.2.1",
     "react-moment": "^0.9.2",
     "react-redux": "^6.0.0",
     "react-router-dom": "^4.3.1",
